Updated README. Dynamically import from teams directory and run scraper on files in that directory.
